compare: Put each group into its own file to be able to run them independently using `cargo bench --bench <group>`
